Assessment 2 - Moved statistics to its own .cs file (as per requirements)
