+++ CMP1903_A1_2324/CMP1903_A1_2324.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Game.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Statistics.cs" />
     <Compile Include="Testing.cs" />
   </ItemGroup>
   <ItemGroup>
